data/common/pyproject.toml
[project]
name = "common"
version = "0.1.0"
description = "Common functions shared across packages."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psycopg2-binary>=2.9.0",
]
